Changes related to: https://github.com/code4moldova/voluntar-backend/pull/100#pullrequestreview-489381131 Closed #91 #92 #93 #90 #89 #88 #84
